+++ jest.config.ts
@@ -5,7 +5,7 @@ const config: Config.InitialOptions = {
   moduleDirectories: ['node_modules', 'lib'],
   testPathIgnorePatterns: ['/node_modules/', 'lib/__tests__/utils'],
   moduleNameMapper: {
-    '^@coast/(.*)$':  '<rootDir>/lib/$1',
+    '^@coast/(.*)$': '<rootDir>/lib/$1',
   },
   transform: {
     '^.+\\.ts$': 'ts-jest',