+++ lib/index.ts
@@ -7,5 +7,14 @@ export { default as CoTabPanel } from './tabs/TabPanel.vue';
 export { default as CoDialog } from './dialog/Dialog.vue';
 export { default as CoCard } from './card/Card.vue';
 export { default as CoIcon } from './icon/Icon.vue';
-export { default as CoInput } from './input/Input.vue';
+export { Toast as CoToast } from './toast';
 export { openDialog as openDialog } from './dialog/openDialog';
+export { default as CoInput } from './input/Input.vue';
+export { Col as CoCol, Row as CoRow } from './grid';
+export {
+  Aside as CoAside,
+  Content as CoContent,
+  Footer as CoFooter,
+  Header as CoHeader,
+  Layout as CoLayout,
+} from './layout';