+++ lib/layout/index.ts
@@ -1,0 +1,6 @@
+import './style/layout.scss';
+export { default as Aside } from './Aside';
+export { default as Content } from './Content';
+export { default as Footer } from './Footer';
+export { default as Header } from './Header';
+export { default as Layout } from './Layout';