+++ src/examples/Toast/index.vue
@@ -1,0 +1,29 @@
+<template>
+  <div>
+    <h1>Toast 文档</h1>
+    <Demo :component="ToastDefaultDemo" />
+    <Demo :component="ToastTypeDemo" />
+    <Demo :component="ToastCenterDemo" />
+    <Demo :component="ToastCloseDemo" />
+  </div>
+</template>
+
+<script lang="ts">
+import ToastDefaultDemo from './Toast.default.demo.vue';
+import ToastTypeDemo from './Toast.type.demo.vue';
+import ToastCenterDemo from './Toast.center.demo.vue';
+import ToastCloseDemo from './Toast.close.demo.vue';
+import Demo from '../../components/Demo.vue';
+
+export default {
+  components: { Demo },
+  setup() {
+    return {
+      ToastDefaultDemo,
+      ToastTypeDemo,
+      ToastCenterDemo,
+      ToastCloseDemo,
+    };
+  },
+};
+</script>